// apis/token-api.ts
export interface TokenApiCreateTokenRequest {
    /**
     * 媒体库ID，在媒体托管控制台创建媒体库后获取。
     */
    readonly libraryId: string

    /**
     * 媒体库密钥，在媒体托管控制台创建媒体库后获取。
     */
    readonly librarySecret: string

    /**
     * 空间ID，可同时指定多个空间ID，使用英文逗号（,）分隔。
     */
    readonly spaceId?: string

    /**
     * 用户身份识别，由业务后台自行控制。
     */
    readonly userId?: string

    /**
     * 客户端识别，由业务后台自行控制。
     */
    readonly clientId?: string

    /**
     * SessionId，由业务后台自行控制。
     */
    readonly sessionId?: string

    /**
     * 令牌有效时长及每次使用令牌后自动续期的有效时长，单位为秒。
     */
    readonly period?: number

    /**
     * 授予的权限，如为空则只授予读取权限。
     */
    readonly grant?: CreateTokenGrantEnum
}

/**
 * Request parameters for deleteToken operation in TokenApi.
 */
export interface TokenApiDeleteTokenRequest {
    /**
     * 媒体库 ID，必选参数
     */
    readonly libraryId: string

    /**
     * 访问令牌
     */
    readonly accessToken: string
}

/**
 * Request parameters for deleteUserTokens operation in TokenApi.
 */
export interface TokenApiDeleteUserTokensRequest {
    /**
     * 媒体库 ID，必选参数
     */
    readonly libraryId: string

    /**
     * 媒体库密钥
     */
    readonly librarySecret: string

    /**
     * 用户身份识别，当访问令牌对应的权限为管理员权限且申请访问令牌时的用户身份识别为空时用来临时指定用户身份，详情请参阅生成访问令牌接口，可选参数
     */
    readonly userId?: string

    /**
     * 客户端识别，多个 ClientId 用英文逗号分隔，一次最多不超过 100 个
     */
    readonly clientId?: string

    /**
     * 会话识别，多个 SessionId 用英文逗号分隔，一次最多不超过 100 个
     */
    readonly sessionId?: string
}

/**
 * Request parameters for renewToken operation in TokenApi.
 */
export interface TokenApiRenewTokenRequest {
    /**
     * 媒体库 ID，必选参数
     */
    readonly libraryId: string

    /**
     * 访问令牌
     */
    readonly accessToken: string
}

/**
 * TokenApi - object-oriented interface
 */
export class TokenApi extends BaseAPI {
    /**
     * 用于生成调用智能媒资托管服务的访问令牌（Access Token）。
     * @summary 生成访问令牌
     * @param {TokenApiCreateTokenRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public createToken(requestParameters: TokenApiCreateTokenRequest, options?: RawAxiosRequestConfig) {
        return TokenApiFp(this.configuration).createToken(requestParameters.libraryId, requestParameters.librarySecret, requestParameters.spaceId, requestParameters.userId, requestParameters.clientId, requestParameters.sessionId, requestParameters.period, requestParameters.grant, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 用于删除指定的访问令牌（Access Token）。删除指定访问令牌无需校验媒体库密钥，故可在客户端调用该接口。
     * @summary 删除访问令牌
     * @param {TokenApiDeleteTokenRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public deleteToken(requestParameters: TokenApiDeleteTokenRequest, options?: RawAxiosRequestConfig) {
        return TokenApiFp(this.configuration).deleteToken(requestParameters.libraryId, requestParameters.accessToken, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 用于删除特定用户的所有访问令牌（Access Token）。调用该接口需要用到媒体库密钥，所以必须在后端调用该接口以保证密钥安全；必须指定 UserId 参数，因此在创建访问令牌时，如果后续计划主动删除对应的访问令牌，则在创建时也需要指定 UserId；
     * @summary 删除特定用户的所有访问令牌
     * @param {TokenApiDeleteUserTokensRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public deleteUserTokens(requestParameters: TokenApiDeleteUserTokensRequest, options?: RawAxiosRequestConfig) {
        return TokenApiFp(this.configuration).deleteUserTokens(requestParameters.libraryId, requestParameters.librarySecret, requestParameters.userId, requestParameters.clientId, requestParameters.sessionId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 用于续期访问令牌（Access Token）。续期时不支持指定新的有效时长，仅按照获取令牌时指定的有效时长续期。
     * @summary 续期访问令牌
     * @param {TokenApiRenewTokenRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public renewToken(requestParameters: TokenApiRenewTokenRequest, options?: RawAxiosRequestConfig) {
        return TokenApiFp(this.configuration).renewToken(requestParameters.libraryId, requestParameters.accessToken, options).then((request) => request(this.axios, this.basePath));
    }
}

export const CreateTokenGrantEnum = {
    Admin: 'admin',
    CreateSpace: 'create_space',
    DeleteSpace: 'delete_space',
    SpaceAdmin: 'space_admin',
    CreateDirectory: 'create_directory',
    DeleteDirectory: 'delete_directory',
    DeleteDirectoryPermanent: 'delete_directory_permanent',
    MoveDirectory: 'move_directory',
    CopyDirectory: 'copy_directory',
    UploadFile: 'upload_file',
    UploadFileForce: 'upload_file_force',
    BeginUpload: 'begin_upload',
    BeginUploadForce: 'begin_upload_force',
    ConfirmUpload: 'confirm_upload',
    CreateSymlink: 'create_symlink',
    CreateSymlinkForce: 'create_symlink_force',
    DeleteFile: 'delete_file',
    DeleteFilePermanent: 'delete_file_permanent',
    MoveFile: 'move_file',
    MoveFileForce: 'move_file_force',
    CopyFile: 'copy_file',
    CopyFileForce: 'copy_file_force',
    DeleteRecycled: 'delete_recycled',
    RestoreRecycled: 'restore_recycled'
} as const;
export type CreateTokenGrantEnum = typeof CreateTokenGrantEnum[keyof typeof CreateTokenGrantEnum];